+++ seleniumpm/examples/google_page.py
@@ -14,6 +14,9 @@ class GooglePage(Webpage):
         self.path = ""
         self.search_field = TextField(driver, Locator(By.NAME, 'q'))
 
+    def get_result_links(self):
+        return self.driver.find_elements(By.XPATH, "//h3[contains(@class, 'r')]/a")
+
     def wait_for_page_load(self, timeout=30):
         """
         This is an overridden method to support checks for successful page load.